@@ -12,28 +12,28 @@
     },
     "license": "ISC",
     "dependencies": {
-        "@angular/animations": "^5.0.0",
-        "@angular/common": "^5.0.0",
-        "@angular/compiler": "^5.0.0",
-        "@angular/core": "^5.0.0",
-        "@angular/forms": "^5.0.0",
-        "@angular/http": "^5.0.0",
-        "@angular/platform-browser": "^5.0.0",
-        "@angular/platform-browser-dynamic": "^5.0.0",
-        "@angular/router": "^5.0.0",
-        "core-js": "^2.5.1",
-        "rxjs": "^5.5.2",
-        "zone.js": "^0.8.18",
-        "systemjs": "^0.20.19"
+        "@angular/animations": "^6.1.6",
+        "@angular/common": "^6.1.6",
+        "@angular/compiler": "^6.1.6",
+        "@angular/core": "^6.1.6",
+        "@angular/forms": "^6.1.6",
+        "@angular/http": "^6.1.6",
+        "@angular/platform-browser": "^6.1.6",
+        "@angular/platform-browser-dynamic": "^6.1.6",
+        "@angular/router": "^6.1.6",
+        "core-js": "^2.5.7",
+        "rxjs": "^6.3.1",
+        "zone.js": "^0.8.26",
+        "systemjs": "^0.21.4"
     },
     "devDependencies": {
-        "@types/node": "~8.0.53",
-        "typescript": "~2.4.2",
+        "@types/node": "~10.9.4",
+        "typescript": "2.9.2",
         "gulp": "^3.9.1",
         "gulp-concat": "^2.6.1",
-        "gulp-plumber": "^1.1.0",
-        "gulp-uglify": "^3.0.0",
-        "run-sequence": "^2.2.0",
+        "gulp-plumber": "^1.2.0",
+        "gulp-uglify": "^3.0.1",
+        "run-sequence": "^2.2.1",
         "del": "^3.0.0",
         "del-cli": "^1.1.0",
         "webpack": "3.8.1",
