@@ -1,6 +1,8 @@
 import { NgModule, Optional, SkipSelf } from '@angular/core';
-import { HttpModule, Request, XSRFStrategy, CookieXSRFStrategy } from '@angular/http';
-import { HttpClientModule } from '@angular/common/http';
+import { HttpClientModule, HttpClientXsrfModule } from '@angular/common/http';
+
+//Can use with Angular 2+ but we're using HttpClientModule (see above) for this app
+//import { HttpModule, XSRFStrategy, CookieXSRFStrategy } from '@angular/http';
 
 import { DataService } from './data.service';
 import { DataFilterService } from './data-filter.service';
@@ -8,11 +10,26 @@ import { Sorter } from './sorter';
 import { TrackByService } from './trackby.service';
 import { EnsureModuleLoadedOnceGuard } from '../shared/ensureModuleLoadedOnceGuard';
 
+/* 
+    ##### PLEASE NOTE ######
+    This code has been updated to use the HttpClientModule that's part of Angular 4.3+
+    Http and HttpModule have been deprecated
+    #####
+*/
+
 @NgModule({
-  //If you're on Angular 4.3 or higher you can use HttpClientModule. See core.module.ts.httpclient
-  imports: [ HttpModule, HttpClientModule ],
+  imports: [ 
+    //Can use with Angular 4.3+ to 
+    HttpClientModule, 
+    //Can use to override default names for XSRF cookie and header
+    // HttpClientXsrfModule.withOptions({
+    //   cookieName: 'My-XSRF-TOKEN',
+    //   headerName: 'My-X-XSRF-TOKEN',
+    // })
+  ],
   providers: [
     //Default XSRF provider setup (change cookie or header name if needed): 
+    //Can use this with Angular 2+ but if using HttpClientModule use options as shown above
     //{ provide: XSRFStrategy, useValue: new CookieXSRFStrategy('XSRF-TOKEN', 'X-XSRF-TOKEN') },
     DataService, DataFilterService, Sorter, TrackByService] // these should be singleton
 })
